Fixed bug related to object pool - disappearing projectiles.
When projectile is thrown, Projectile script sets timer after which that
projectile will be released (in case it won't hit anything). But when it
hits some collider, it is released too. So when a projectile is thrown,
then it collides something and dissapears in time smaller than
max life time of that projectile, and after that player will shoot
another projectile (which, thanks to ObjectPool, it can be the same
projectile), the coroutine releasing the object set earlier, can release
it in that second shot. Then You can see sometimes a sisappearing
projectiles.